Get and transform natural earth states data
Since Times Magazine does not include geographic data
for the election results, we need to get the gis data
to join with the existing election results.

This commit adds a reproducible script for getting,
filtering, and simplifying natural earth state data
so we only have the US states (and only relevant)
fields to be joined to election results.