Hardcode SVG dimensions
Ran into an error when trying to calculate the
DOM dimensions dynamically with this.$el.getBoundingClientRect().

Seems to be an error of operations order in the lifecycle,
that only happens sometimes (when computing the state path
with d3).

Hardcoding the height and width values for now so I can go
back when I have more time to investigate this strange bug.