Render static map; fix async mapData issue
This commit fixes the problem that I was encountering
where the computed mapData was trying to access state.features
before the state geojson has loaded into Vuex store.

Added a check to only run the d3 math calculations for state paths
if state.features exists (testing to ensure data is loaded).

Vue's computed properties are reactive, so mapData() will be
rerun once the computed state() data is available (so state.features)
is not undefined.

This was also the issue the component was facing with dynamically
defined width and height, so I switched back from the hard-coded
temporary fix.

See here https://stackoverflow.com/questions/52121504/computed-property-needs-to-wait-for-async-data